etcd-runner: refactor code structure and flag cleanup
